Modify agent command
